+++ RELEASES.md
@@ -90,6 +90,7 @@ Upcoming
 - Extend functionality of GetTibcoQueues (currently used in "ShowTibcoQueues" in IJA_TiBeT2)
 - Remove useless space in SOAP envelop element
 - Improve code to prevent double attributes in merged schema
+- Add useCdataSection attribute to Text2XmlPipe